Support pre-defined answers for `./configure`
If ran regularly, it might be useful for the `configure` program to
accept pre-defined answers to the interactive questions.

This commit adds support for eight environment variables to the
`./configure` program.

For seven of those eight variables, it will check whether they are
set; if so, it will skip the interactive question for this variable.

The `SMT_DATE_FORMAT_STRING` variable is a special case; it will make
the `./configure` program use the format provided in this variable
but it will still ask for today’s date interactively.

This adds support for a total of eight environment variables:

- SMT_PROJECT_NAME
- SMT_ORGANIZATION_NAME
- SMT_COM_AN_ORGANIZATION_IDENTIFIER
- SMT_AUTHOR_NAME
- SMT_TODAYS_DATE
- SMT_TODAYS_YEAR
- SMT_GITHUB_USERNAME
- SMT_DATE_FORMAT_STRING

This commit also adds a section called *Setting up pre-defined
answers* to `README.md`.