Automate search-replace steps in recipe
This commit tries to leverage the (macOS flavor of) `find` and `sed`
to automate steps 5.1.1 and 5.1.2. of the recipe, ultimately merging
both into a single step of 5.1.

Hopefully, this saves template maintainers a bit of time when
executing those recipe steps and is less tedious. It also loses the
dependency on the text editor.