Remove references to development team ID (#30)
This commit removes all references to development team IDs from
`project.pbxproj` files, which would lead to unsightly warnings if
that particular team ID is not registered with Xcode.

The recipe is also updated to reflect this change. Since the recipe
already assumes we’re on macOS, we can use (the macOS flavors of)
`find` and `sed` to change all occurrences in place, which has the
benefit that no text editor is needed while executing that step in
the recipe.